# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/pin/status:
    post:
      tags: [Admin]
      operationId: getAdminPinStatus
      summary: Whether the account has an admin PIN, and its lockout state
      description: |
        Called when the user opens the admin panel so the app can decide
        between the "create PIN", "enter PIN" and "locked" flows.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPinStatusRequest"
      responses:
        "200":
          description: PIN status for the account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPinStatusResult"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/pin/setup:
    post:
      tags: [Admin]
      operationId: setupAdminPin
      summary: Create or replace the account's admin PIN
      description: |
        Used for the first-time PIN and for "I forgot my PIN" (after the app
        has re-verified the account password client-side). Resets the failed
        attempt counter and any lockout, and returns a fresh admin session
        token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPinSetupRequest"
      responses:
        "200":
          description: PIN stored; admin session issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPinSession"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/pin/verify:
    post:
      tags: [Admin]
      operationId: verifyAdminPin
      summary: Verify the account's admin PIN
      description: |
        On success returns an admin session token and resets the failed
        attempt counter. On failure increments it; after 5 consecutive
        failures the PIN is locked for 15 minutes and `lockedUntil` is set.
        While locked, no PIN is checked.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPinVerifyRequest"
      responses:
        "200":
          description: |
            Verification result. `ok` true carries a `token`; `ok` false
            carries `remainingAttempts` and, when locked, `lockedUntil`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPinVerifyResult"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/pin/session:
    post:
      tags: [Admin]
      operationId: createAdminSession
      summary: Issue an admin session without a PIN (biometric / password path)
      description: |
        Called after the app has verified the user locally — either a device
        biometric check, or (while the PIN is locked) re-entering the account
        password. Resets the failed attempt counter and lockout, and returns
        an admin session token. There is no server-side proof of the local
        check, matching WODPLACE's current trusted-client model.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSessionRequest"
      responses:
        "200":
          description: Admin session issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminPinSession"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: No admin PIN exists for this account yet.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/contracts:
    get:
      tags: [Admin]
      operationId: listAdminContracts
      summary: List contract documents for the admin panel
      responses:
        "200":
          description: Contract documents.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContractDocument"
        "401":
          description: Missing or invalid admin session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/contracts/{slug}:
    put:
      tags: [Admin]
      operationId: updateAdminContract
      summary: Attach a newly uploaded PDF to a contract document slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminContractRequest"
      responses:
        "200":
          description: Updated document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDocument"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing or invalid admin session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/contract-acceptances:
    get:
      tags: [Admin]
      operationId: listAdminContractAcceptances
      summary: List contract acceptances for the owner notification badge
      description: |
        Returns every recorded contract acceptance (newest first) so the
        admin panel can show a badge/counter for ones the owner hasn't
        viewed yet (`seen: false`).
      responses:
        "200":
          description: Contract acceptances.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContractAcceptanceNotification"
        "401":
          description: Missing or invalid admin session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/contract-acceptances/ack:
    post:
      tags: [Admin]
      operationId: ackContractAcceptances
      summary: Mark all pending contract acceptances as seen by the owner
      description: |
        Called when the owner opens the notifications section of the admin
        panel. Clears the unseen badge/counter.
      responses:
        "200":
          description: Number of acceptances that were marked seen.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AckContractAcceptancesResult"
        "401":
          description: Missing or invalid admin session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/dash-link:
    post:
      tags: [Admin]
      operationId: createAdminDashLink
      summary: Mint a one-time auto-login link for the box or super-admin dashboard
      description: |
        Given a valid admin session token, resolves the caller's Supabase
        Auth account (via `boxes.owner_user_id`, falling back to an email
        match against `profiles` that also carries a `box_admin` /
        `super_admin` role) and returns a single-use Supabase magic link that
        logs that account into the requested panel (`target`, default
        "box"). The link is short-lived and must be opened immediately.
        Returns 409 when no linked Supabase admin account can be resolved —
        the caller should then load that panel's normal email/password
        login. Returns 403 when `target: "super"` is requested by an account
        without the super_admin role.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminDashLinkRequest"
      responses:
        "200":
          description: One-time auto-login URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashLinkResult"
        "401":
          description: Missing or invalid admin session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: "target: \"super\" requested by a non-super_admin account."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: No linked Supabase admin account for this PIN identity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "502":
          description: Upstream Supabase call failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ContractDocument:
      type: object
      required: [slug, title, read]
      properties:
        slug:
          type: string
        title:
          type: string
        objectPath:
          type: string
          nullable: true
        read:
          type: boolean
        readAt:
          type: string
          nullable: true
...
    AdminPinStatusRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          minLength: 1
    AdminPinStatusResult:
      type: object
      required: [hasPin, failedAttempts, lockedUntil]
      properties:
        hasPin:
          type: boolean
        failedAttempts:
          type: integer
        lockedUntil:
          type: string
          nullable: true
          description: ISO timestamp until which PIN entry is locked, or null.
    AdminPinSetupRequest:
      type: object
      required: [userId, pin]
      properties:
        userId:
          type: string
          minLength: 1
        pin:
          type: string
          pattern: '^[0-9]{4,6}$'
          description: 4 to 6 digits.
    AdminPinVerifyRequest:
      type: object
      required: [userId, pin]
      properties:
        userId:
          type: string
          minLength: 1
        pin:
          type: string
          minLength: 1
    AdminPinVerifyResult:
      type: object
      required: [ok, remainingAttempts, lockedUntil]
      properties:
        ok:
          type: boolean
        token:
          type: string
          nullable: true
          description: Admin session token; present only when `ok` is true.
        remainingAttempts:
          type: integer
          description: PIN attempts left before the 15-minute lockout.
        lockedUntil:
          type: string
          nullable: true
          description: ISO timestamp until which PIN entry is locked, or null.
    AdminSessionRequest:
      type: object
      required: [userId]
      properties:
        userId:
          type: string
          minLength: 1
    AdminPinSession:
      type: object
      required: [ok, token]
      properties:
        ok:
          type: boolean
        token:
          type: string
          description: Admin session token, sent as an Authorization bearer header.
    UpdateAdminContractRequest:
      type: object
      required: [objectPath]
      properties:
        objectPath:
          type: string
          minLength: 1
        title:
          type: string
    ContractAcceptanceNotification:
      type: object
      required:
        - userId
        - name
        - email
        - emergencyContactName
        - emergencyContactPhone
        - acceptedAt
        - seen
      properties:
        userId:
          type: string
        name:
          type: string
        email:
          type: string
        emergencyContactName:
          type: string
        emergencyContactPhone:
          type: string
        acceptedAt:
          type: string
        guardianName:
          type: string
          nullable: true
        guardianRelationship:
          type: string
          nullable: true
        seen:
          type: boolean
    AckContractAcceptancesResult:
      type: object
      required: [acknowledged]
      properties:
        acknowledged:
          type: integer
    AdminDashLinkResult:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: |
            Single-use Supabase magic link. Open it in a WebView
            immediately; it expires quickly and is consumed on first use.
    CreateAdminDashLinkRequest:
      type: object
      properties:
        target:
          type: string
          enum: [box, super]
          default: box
          description: |
            Which panel to mint the link for. "box" (default, preserves the
            existing behavior) redirects into box-admin; "super" redirects
            into the super-admin-hub and is rejected with 403 unless the
            caller's resolved admin roles include super_admin.
...